jeff_agent/pyproject.toml
[project]
name = "jeff-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.23",
    "google-adk>=1.25.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "python-dotenv>=1.2.1",
    "sse-starlette>=3.2.0",
    "starlette>=0.52.1",
    "uvicorn>=0.41.0",
]
